# explicitly using int64 so it works consistently